qdts_client/pyproject.toml
[tool.poetry]
name = "qdts_client"
version = "0.2.0"
description = "Client lib for qdts"
authors = ["rmcelaya <rmcelaya@protonmail.com>"]

[tool.poetry.dependencies]
protobuf = "*"
grpcio = "*"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
